requirements.txt
streamlit
pdfplumber
python-docx
google-generativeai
PyPDF2
pdfminer.six
regex
